requirements.txt
altgraph==0.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.0.1
pydantic
supabase
streamlit
resend

 
certifi==2025.11.12
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
future==0.18.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
macholib==1.15.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pandas==2.3.3
pip==25.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.48.0
setuptools==58.0.4
shellingham==1.5.4
six==1.16.0
SQLAlchemy==2.0.45
starlette==0.49.3
streamlit==1.53.0
extra-streamlit-components==0.1.71
supabase==2.27.2
tomli==2.3.0
typer==0.21.0
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
uvicorn==0.39.0
watchfiles==1.1.1
websockets==15.0.1
wheel==0.37.0
plotly